requirements.txt
Django==1.10.2
django-redis==4.5.0
djangorestframework==3.4.7
psycopg2==2.6.2
redis==2.10.5
requests==2.11.1
